camera: Fix initial camera position (should still set theta and phi aswell), fix new center position, specify texcoords before vertices in RaceTrack, refactor Robot + start on drawing sphere head

+++ src/robotrace/Terrain.java
@@ -15,8 +15,8 @@ class Terrain {
     private final int minY = -20;
     private final int maxY = 20;    
     
-    private int xInterval;
-    private int yInterval;
+    private float xInterval;
+    private float yInterval;
     
     public Terrain() {
         xInterval = (maxX - minX) / fragments;
